customers/elkjøp/div4c/shield trunk
Fix various issues with group by clauses.

Group by together with wildcards was unreliable and failed in lots of interesting and subtle ways on complicated queries. This patch cleans up the code significantly, separates the group by handling into a set of separate steps and moves a lot of the work to a new catalyst.

git-svn-id: https://projects.freecode.no/internal/svn/customers/elkj%C3%B8p/div4c@397 3f367b02-151f-0410-a67f-93a54fc96dc6