customers/elkjøp/div4c/shield trunk
Fix two weird Oracle issues, not allowing null tests in field output and failing to convert dates in some selects.

This patch fixes two Oracle issues. The first, as near as I can tell is an oracle bug. If calling something like 'date_format(agg (foo))', where 'date_format' is a function using to_char to convert a date to a string and 'agg' is a group aggregation function, the date function will fail if the grouping is performed inside a subselect. 

The second issue is that null tests like 'is null' are not allowed as field output in selects.

git-svn-id: https://projects.freecode.no/internal/svn/customers/elkj%C3%B8p/div4c@412 3f367b02-151f-0410-a67f-93a54fc96dc6