customers/elkjøp/ecco/div4c/ecco trunk
Fix bug in limit clauses causing incorrect returned values.

Oracle is unfathomably stupid in how hard it makes it to implement limit-like functionality. It turns out that you actually need to have a double subselect with a very specific syntax to implement this so that it will work correctly with order-clauses.

git-svn-id: https://projects.freecode.no/internal/svn/customers/elkj%C3%B8p/div4c@552 3f367b02-151f-0410-a67f-93a54fc96dc6